:recycle: プロトタイプの実装処理とプロトタイプのテスト挙動処理を分割
